+++ delete.ps1
@@ -1,7 +1,7 @@
-#################################################
-# HelloID-Conn-Prov-Target-Nedap-AEOS-Enable
+##################################################
+# HelloID-Conn-Prov-Target-Nedap-AEOS-Delete
 # PowerShell V2
-#################################################
+##################################################
 
 # Enable TLS1.2
 [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor [System.Net.SecurityProtocolType]::Tls12
@@ -41,39 +41,6 @@ function Resolve-NedapAEOSError {
         }
         Write-Output $httpErrorObj
     }
-}
-
-function New-SoapBodyEnableEmployee {
-    [CmdletBinding()]
-    param (
-        [Parameter(Mandatory)]
-        [string]$Id
-    )
-    $now = (Get-Date).ToUniversalTime()
-    $dateFormat = "yyyy-MM-ddTHH:mm:ss"
-
-    $root = "sch:EmployeeChange"
-    $body = "  <sch:Id>$Id</sch:Id>`n"
-    $body += "  <sch:ArrivalDateTime>$($now.ToString($dateFormat))</sch:ArrivalDateTime>`n"
-    $body += "  <sch:LeaveDateTime>2099-01-01T00:00:00</sch:LeaveDateTime>"
-
-    Write-Output "<$root>`n$body`n</$root>"
-}
-
-function New-SoapBodyFindEmployeeById {
-    [CmdletBinding()]
-    param (
-        [Parameter(Mandatory)]
-        [string]$Id
-    )
-    [system.text.StringBuilder]$soapFindEmployee = [system.text.StringBuilder]::new()
-    $null = $soapFindEmployee.Append('<sch:EmployeeSearchInfo>')
-    $null = $soapFindEmployee.Append('<sch:EmployeeInfo>')
-    $null = $soapFindEmployee.Append("<sch:Id>$Id</sch:Id>")
-    $null = $soapFindEmployee.Append('</sch:EmployeeInfo>')
-    $null = $soapFindEmployee.Append('</sch:EmployeeSearchInfo>')
-
-    Write-Output $soapFindEmployee.ToString()
 }
 
 function Invoke-NedapAEOSRestMethod {
@@ -121,6 +88,58 @@ function Invoke-NedapAEOSRestMethod {
 }
 
 
+function New-SoapBodyDisableEmployee {
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory)]
+        [string]$Id
+    )
+    $now = (Get-Date).ToUniversalTime()
+    $dateFormat = "yyyy-MM-ddTHH:mm:ss"
+
+    $root = "sch:EmployeeChange"
+    $body = "  <sch:Id>$Id</sch:Id>`n"
+    $body += "  <sch:LeaveDateTime>$($now.ToString($dateFormat))</sch:LeaveDateTime>"
+
+    Write-Output "<$root>`n$body`n</$root>"
+}
+
+function New-SoapBodyFindCarrierToken {
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory)]
+        [string]$CarrierId
+    )
+    $root = "sch:CarrierIdToken"
+    Write-Output "<$root>$CarrierId</$root>"
+}
+
+function New-SoapBodyFindEmployeeById {
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory)]
+        [string]$Id
+    )
+    [system.text.StringBuilder]$soapFindEmployee = [system.text.StringBuilder]::new()
+    $null = $soapFindEmployee.Append('<sch:EmployeeSearchInfo>')
+    $null = $soapFindEmployee.Append('<sch:EmployeeInfo>')
+    $null = $soapFindEmployee.Append("<sch:Id>$Id</sch:Id>")
+    $null = $soapFindEmployee.Append('</sch:EmployeeInfo>')
+    $null = $soapFindEmployee.Append('</sch:EmployeeSearchInfo>')
+
+    Write-Output $soapFindEmployee.ToString()
+}
+
+
+function New-SoapBodyWithdrawCarrierToken {
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory)]
+        [PSCustomObject]$Identifier
+    )
+    $root = "sch:IdentifierWithdraw"
+    Write-Output "<$root><sch:IdentifierType>$($Identifier.IdentifierType)</sch:IdentifierType><sch:BadgeNumber>$($Identifier.BadgeNumber)</sch:BadgeNumber></$root>"
+}
 #endregion
 
 try {
@@ -132,13 +151,14 @@ try {
     # Setup credentials
     $securePassword = $actionContext.Configuration.Password | ConvertTo-SecureString -AsPlainText -Force
     $credential = [System.Management.Automation.PSCredential]::new($actionContext.Configuration.UserName, $securePassword)
+
     Write-Information 'Verifying if a Nedap-AEOS account exists'
     $soapBody = New-SoapBodyFindEmployeeById -Id $actionContext.References.Account
     $response = Invoke-NedapAEOSRestMethod -Uri $actionContext.Configuration.BaseUrl -SoapBody $soapBody -Credential $credential
     $correlatedAccount = $response.Envelope.Body.EmployeeList.Employee.EmployeeInfo
 
     if ($null -ne $correlatedAccount) {
-        $action = 'EnableAccount'
+        $action = 'DisableAccount'
     } 
     else {
         $action = 'NotFound'
@@ -146,20 +166,36 @@ try {
 
     # Process
     switch ($action) {
-        'EnableAccount' {
+        'DisableAccount' {
             if (-not($actionContext.DryRun -eq $true)) {
-                Write-Information "Enabling Nedap-AEOS account with accountReference: [$($actionContext.References.Account)]"
+                Write-Information "Disabling Nedap-AEOS account with accountReference: [$($actionContext.References.Account)]"
 
-                $soapBody = New-SoapBodyEnableEmployee -Id $actionContext.References.Account
+                # Withdraw all carrier tokens (badges) from the employee
+                $soapBody = New-SoapBodyFindCarrierToken -CarrierId $correlatedAccount.Id
+                $response = Invoke-NedapAEOSRestMethod -Uri $actionContext.Configuration.BaseUrl -SoapBody $soapBody -Credential $credential
+
+                $identifier = $null
+                if ($null -ne $Response.Envelope.Body) {
+                    foreach ($Item in  $Response.Envelope.Body.IdentifierList.Identifier) {
+                        $identifier = $Item
+                        if ($null -ne $identifier) {
+                            $soapBody = New-SoapBodyWithdrawCarrierToken -Identifier $identifier
+                            $response = Invoke-NedapAEOSRestMethod -Uri $actionContext.Configuration.BaseUrl -SoapBody $soapBody -Credential $credential
+                        }
+                    }
+                }
+
+                # Disable the employee by setting LeaveDateTime to current time
+                $soapBody = New-SoapBodyDisableEmployee -Id $actionContext.References.Account
                 $response = Invoke-NedapAEOSRestMethod -Uri $actionContext.Configuration.BaseUrl -SoapBody $soapBody -Credential $credential
             } 
             else {
-                Write-Information "[DryRun] Enable Nedap-AEOS account with accountReference: [$($actionContext.References.Account)], will be executed during enforcement"
+                Write-Information "[DryRun] Disable Nedap-AEOS account with accountReference: [$($actionContext.References.Account)], will be executed during enforcement"
             }
 
             $outputContext.Success = $true
             $outputContext.AuditLogs.Add([PSCustomObject]@{
-                    Message = 'Enable account was successful'
+                    Message = "Disable account [$($actionContext.References.Account)] was successful. Action initiated by: [$($actionContext.Origin)]"
                     IsError = $false
                 })
             break
@@ -167,15 +203,14 @@ try {
 
         'NotFound' {
             Write-Information "Nedap-AEOS account: [$($actionContext.References.Account)] could not be found, indicating that it may have been deleted"
-            $outputContext.Success = $false
+            $outputContext.Success = $true
             $outputContext.AuditLogs.Add([PSCustomObject]@{
-                    Message = "Nedap-AEOS account: [$($actionContext.References.Account)] could not be found, indicating that it may have been deleted"
-                    IsError = $true
+                    Message = "Nedap-AEOS account: [$($actionContext.References.Account)] could not be found, indicating that it may have been deleted. Action initiated by: [$($actionContext.Origin)]"
+                    IsError = $false
                 })
             break
         }
     }
-
 } 
 catch {
     $outputContext.success = $false
@@ -183,11 +218,11 @@ catch {
     if ($($ex.Exception.GetType().FullName -eq 'Microsoft.PowerShell.Commands.HttpResponseException') -or
         $($ex.Exception.GetType().FullName -eq 'System.Net.WebException')) {
         $errorObj = Resolve-NedapAEOSError -ErrorObject $ex
-        $auditLogMessage = "Could not enable Nedap-AEOS account. Error: $($errorObj.FriendlyMessage)"
+        $auditLogMessage = "Could not disable Nedap-AEOS account. Error: $($errorObj.FriendlyMessage). Action initiated by: [$($actionContext.Origin)]"
         Write-Warning "Error at Line '$($errorObj.ScriptLineNumber)': $($errorObj.Line). Error: $($errorObj.ErrorDetails)"
     } 
     else {
-        $auditLogMessage = "Could not enable Nedap-AEOS account. Error: $($ex.Exception.Message)"
+        $auditLogMessage = "Could not disable Nedap-AEOS account. Error: $($ex.Exception.Message). Action initiated by: [$($actionContext.Origin)]"
         Write-Warning "Error at Line '$($ex.InvocationInfo.ScriptLineNumber)': $($ex.InvocationInfo.Line). Error: $($ex.Exception.Message)"
     }
     $outputContext.AuditLogs.Add([PSCustomObject]@{